Substrate refactor and connection issues (#189)
* substrate refactor and restoring connection to substrate and centrifuge

* Fixing comestic issues on connection and using alice account by default on connection

* Fixing issue with file, reverting to original style

* Moving reducer logic to custom substrate reducer

* Fixing errors on reducer and declaring new type for substrate reducer

* Removing logs
