+++ src/contexts/Adaptors/SubstrateAdaptors/index.ts
@@ -1,0 +1,8 @@
+export { default as SubstrateDestinationAdaptorProvider } from "./SubstrateDestinationAdaptorProvider";
+export { default as SubstrateHomeAdaptorProvider } from "./SubstrateHomeAdaptorProvider";
+export { default as getRelayerThresholdFunc } from "./getRelayerThreshold";
+export { default as confirmChainIdFunc } from "./confirmChainId";
+export { default as queryData } from "./queryData";
+export { default as handleConnectFunc } from "./handleConnect";
+export { default as makeDeposit } from "./makeDeposit";
+export { default as unsubscribeFunc } from "./unsuscribe";
