+++ src/contexts/NetworkManagerContext/NetworkManagerContext.tsx
@@ -183,7 +183,7 @@ const NetworkManagerProvider = ({ children }: INetworkManagerProviderProps) => {
   const [depositNonce, setDepositNonce] = useState<string | undefined>(
     undefined
   );
-  // const [depositVotes, setDepositVotes] = useState<number>(0);
+  const [depositVotes, setDepositVotes] = useState<number>(0);
   const [inTransitMessages, tokensDispatch] = useReducer(
     transitMessageReducer,
     { txIsDone: false, transitMessage: [] }
